06/03/2024
implemented confetti and a partyMode utilizing internal system clock functions for a one-second delay between background changes. PartyMode is toggled on and off by pressing the 'P' key, and does not have a significant impact on system functions.

Additionally, confetti is created using a similar array float function as the fireworks.